Check for device removal during scan.
If a device is removed during a scan, there's a potential timing hole where a
scan will determine that all remaining directories have been deleted before the
removal notification is handled. To correct this, use the MountInfo attached to
the directory to determine if the device is removable. If it is removalble, test
the existance of the directory before updating the delete list for a subirectory.