Provide device mount information to directories.
Add a MountInfo structure that contains information about the mount. It currently
contains a flag that indicates that the device is removable so that a user of the
directory can be prepared for it to become invalid.

In the future, this can include the mount point, directives to use relative
paths, etc.